+++ packages/auth/src/utils/AuthNetworkType.ts
@@ -1,4 +1,5 @@
 export enum AuthNetworkType {
   EVM = 'evm',
   SOLANA = 'solana',
+  APTOS = 'aptos',
 }