+++ packages/common/core/src/Modules/Module.test.ts
@@ -12,12 +12,6 @@ describe('Module', () => {
     public start(): void | Promise<void> {
       // Nothing
     }
-    public getListenerCount(eventName: string): number {
-      return this.emitter.listenerCount(eventName);
-    }
-    public emit(eventName: string, value: any) {
-      this.emitter.emit(eventName, value);
-    }
   }
 
   let core: Core;
@@ -36,27 +30,6 @@ describe('Module', () => {
 
   it('should be able to call start', () => {
     module.start();
-  });
-
-  it('should listen to events', () => {
-    const EVENT_NAME = 'TestEvent';
-    let called = 0;
-
-    const cleanup = module.listen(EVENT_NAME, (value: string) => {
-      expect(value).toBe('success');
-      called++;
-    });
-
-    expect(module.getListenerCount(EVENT_NAME)).toEqual(1);
-    expect(called).toEqual(0);
-
-    module.emit(EVENT_NAME, 'success');
-
-    expect(called).toEqual(1);
-
-    cleanup();
-
-    expect(module.getListenerCount(EVENT_NAME)).toEqual(0);
   });
 
   it('should cleanup the class correctly', () => {