+++ packages/next/src/hooks/auth/index.ts
@@ -1,2 +1,3 @@
 export * from './useAuthRequestChallengeEvm';
 export * from './useAuthRequestChallengeSolana';
+export * from './useAuthRequestChallengeAptos';