+++ packages/react/src/context/index.ts
@@ -1,0 +1,2 @@
+export { default as MoralisProvider } from './MoralisProvider';
+export * from './types';