+++ packages/client/firebaseEvmApi/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @moralisweb3/client-firebase-evm-api
 
+## 2.14.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/client-evm-api@2.14.1
+  - @moralisweb3/client-firebase-api-utils@2.14.1
+  - @moralisweb3/common-core@2.14.1
+  - @moralisweb3/common-evm-utils@2.14.1
+
+## 2.14.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/client-evm-api@2.14.0
+  - @moralisweb3/client-firebase-api-utils@2.14.0
+  - @moralisweb3/common-core@2.14.0
+  - @moralisweb3/common-evm-utils@2.14.0
+
 ## 2.13.0
 
 ### Patch Changes